Completed the class that loads the JSON config file

+++ source/src/main/java/invaders/filehandler/PlayerConfigReader.java
@@ -10,7 +10,7 @@ import org.json.simple.parser.*;
 public interface PlayerConfigReader {
     public JSONObject getPlayerConfig();
     public String getPlayerColour();
-    public int getPlayerSpeed();
-    public int getPlayerLives();
+    public Long getPlayerSpeed();
+    public Long getPlayerLives();
     public Coordinates getPlayerCoordinates();
 }
